+++ packages/liveblocks-client/src/client.ts
@@ -29,13 +29,6 @@ type EnterOptions<
      * https://liveblocks.io/docs/guides/troubleshooting#stale-props-zombie-child
      */
     unstable_batchedUpdates?: (cb: () => void) => void;
-
-    /**
-     * INTERNAL OPTION: Only used in a SSR context when you want an empty room
-     * to make sure your react tree is rendered properly without connecting to
-     * websocket
-     */
-    DO_NOT_USE_withoutConnecting?: boolean;
   }
 >;
 
@@ -99,6 +92,11 @@ export function createClient(options: ClientOptions): Client {
     roomId: string,
     options: EnterOptions<TPresence, TStorage>
   ): Room<TPresence, TStorage, TUserMeta, TRoomEvent> {
+    const shouldConnect =
+      options.shouldInitiallyConnect === undefined
+        ? true
+        : options.shouldInitiallyConnect;
+
     let internalRoom = rooms.get(roomId) as
       | InternalRoom<TPresence, TStorage, TUserMeta, TRoomEvent>
       | undefined;
@@ -148,7 +146,7 @@ export function createClient(options: ClientOptions): Client {
         Json
       >
     );
-    if (!options.DO_NOT_USE_withoutConnecting) {
+    if (shouldConnect) {
       // we need to check here because nextjs would fail earlier with Node < 16
       if (typeof atob === "undefined") {
         if (clientOptions.polyfills?.atob === undefined) {