+++ packages/liveblocks-react/src/factory.tsx
@@ -134,6 +134,7 @@ export function createRoomContext<
       initialPresence,
       initialStorage,
       unstable_batchedUpdates,
+      shouldInitiallyConnect,
     } = props;
 
     if (process.env.NODE_ENV !== "production") {
@@ -165,6 +166,10 @@ export function createRoomContext<
       initialPresence,
       initialStorage,
       unstable_batchedUpdates,
+      shouldInitiallyConnect:
+        shouldInitiallyConnect === undefined
+          ? typeof window !== "undefined"
+          : shouldInitiallyConnect,
     });
 
     const [room, setRoom] = React.useState<
@@ -174,7 +179,7 @@ export function createRoomContext<
         initialPresence: frozen.initialPresence,
         initialStorage: frozen.initialStorage,
         unstable_batchedUpdates: frozen.unstable_batchedUpdates,
-        DO_NOT_USE_withoutConnecting: typeof window === "undefined",
+        shouldInitiallyConnect: frozen.shouldInitiallyConnect,
       } as RoomInitializers<TPresence, TStorage>)
     );
 
@@ -184,7 +189,7 @@ export function createRoomContext<
           initialPresence: frozen.initialPresence,
           initialStorage: frozen.initialStorage,
           unstable_batchedUpdates: frozen.unstable_batchedUpdates,
-          DO_NOT_USE_withoutConnecting: typeof window === "undefined",
+          withoutConnecting: frozen.shouldInitiallyConnect,
         } as RoomInitializers<TPresence, TStorage>)
       );