+++ packages/liveblocks-react/src/types.ts
@@ -28,6 +28,16 @@ export type RoomProviderProps<
     children: React.ReactNode;
 
     /**
+     * Whether or not the room should connect to Liveblocks servers
+     * when the RoomProvider is rendered.
+     *
+     * By default equals to `typeof window !== "undefined"`,
+     * meaning the RoomProvider tries to connect to Liveblocks servers
+     * only on the client side.
+     */
+    shouldInitiallyConnect?: boolean;
+
+    /**
      * If you're on React 17 or lower, pass in a reference to
      * `ReactDOM.unstable_batchedUpdates` or
      * `ReactNative.unstable_batchedUpdates` here.